Create CommonJS dist
