+++ bin/index.js
@@ -1,0 +1,2 @@
+!function(t,e){for(var r in e)t[r]=e[r]}(exports,function(t){var e={};function r(n){if(e[n])return e[n].exports;var o=e[n]={i:n,l:!1,exports:{}};return t[n].call(o.exports,o,o.exports,r),o.l=!0,o.exports}return r.m=t,r.c=e,r.d=function(t,e,n){r.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:n})},r.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},r.t=function(t,e){if(1&e&&(t=r(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var n=Object.create(null);if(r.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var o in t)r.d(n,o,function(e){return t[e]}.bind(null,o));return n},r.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return r.d(e,"a",e),e},r.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},r.p="",r(r.s=23)}([function(t,e){t.exports=require("tslib")},function(t,e){t.exports=require("react")},function(t,e){t.exports=require("react-hot-loader")},function(t,e){t.exports=require("styled-components")},function(t,e){t.exports=function(t){if(!t.webpackPolyfill){var e=Object.create(t);e.children||(e.children=[]),Object.defineProperty(e,"loaded",{enumerable:!0,get:function(){return e.l}}),Object.defineProperty(e,"id",{enumerable:!0,get:function(){return e.i}}),Object.defineProperty(e,"exports",{enumerable:!0}),e.webpackPolyfill=1}return e}},function(t,e,r){"use strict";(function(t){r.d(e,"b",function(){return s}),r.d(e,"a",function(){return c}),r.d(e,"c",function(){return d}),r.d(e,"d",function(){return u});var n,o=r(0);function i(t){return(i="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}(n=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).enterModule)&&n(t);var s;"undefined"!=typeof reactHotLoaderGlobal&&reactHotLoaderGlobal.default.signature;!function(t){t.STRING="string",t.NUMBER="number",t.BOOLEAN="boolean",t.OBJECT="object",t.ARRAY="array",t.ROOT="root"}(s||(s={}));var a,l,c=function(){function t(t,e,r,n,o,i){this.state=null,this.tag=null,this.schemaNode=r,this.config=o,this.resolver=i,this.children=[],this.type=e,this.parentNode=n,this.uid=o.uidGenerator()}return t.prototype.getHandler=function(){return this.handler},t.prototype.isOutputAvailable=function(){return!0},t.prototype.validateCustom=function(){if(0===this.children.length)return[];var t=[];return this.children.forEach(function(e){t=t.concat(e.validateCustom())}),t},t.prototype.getDebugRepresentation=function(){var t="";return t+=this.type.toString(),this.tag&&(t+="["+this.tag+"]"),this.children.length>0&&(t+=" {\n",this.children.map(function(e){var r=e.getDebugRepresentation();t+="  > ",t+=r.split("\n").map(function(t,e){return 0===e?t:" ".repeat(4)+t}).join("\n"),t+="\n"}),t+="}"),t},t.prototype.getUID=function(){return this.uid},t.prototype.onResolveFinished=function(){},t.prototype.modifyContext=function(t){this.config.rootModifyContext(t,this)},t.prototype.setContext=function(t){this.config.rootSetContext(t,this)},t.prototype.resolve=function(){this.state=this.resolveInitialState()||null,this.children=this.resolveChildren()||[],this.onResolveFinished()},t.prototype.addChild=function(t){this.children.push(t)},t.prototype.overrideChildren=function(t){this.children=t.slice()},t.prototype.getChildren=function(){return this.children},t.prototype.resolveNode=function(t,e,r){return this.resolver(t,e,r)},t.prototype.getConfig=function(){return this.config},t.prototype.getSchema=function(){return this.schemaNode},t.prototype.getState=function(){return this.state},t.prototype.getTag=function(){return this.tag},t.prototype.setTag=function(t){this.tag=t},t.prototype.setParent=function(t){this.parentNode=t},t.prototype.getRawOutput=function(t){return null},t.prototype.onChildStateChanged=function(t,e,r){},t.prototype.setStateSilently=function(t,e,r){t&&Object.assign(this.state,o.__assign({},this.state,t))},t.prototype.setState=function(t,e,r){this.setStateSilently(t,e,r),this.parentNode.onChildStateChanged(this.state,this,r)},t.prototype.findChild=function(t){return this.children.filter(function(e){return e.tag===t})[0]},t.prototype.resolveInitialState=function(){return null},t.prototype.resolveChildren=function(){return[]},t.prototype.getOutput=function(t){return(!t||t&&!1!==t.enableFormat)&&this.getSchema().formatOutput?{__data:this.getSchema().formatOutput(this.getRawOutput(t)),__source:this}:{__data:this.getRawOutput(t),__source:this}},t.prototype.setValue=function(t){},t}(),d=function(t){function e(e,r,n,o,i){var a=t.call(this,null,s.ROOT,e,null,r,n)||this;return a.dataTransformer=o,a.validator=i,a}return o.__extends(e,t),e.prototype.resolveInitialState=function(){return{}},e.prototype.validate=function(){this.validator(this)},e.prototype.getOutput=function(t){return 0===this.getChildren().length?{__data:null,__source:this}:1===this.getChildren().length?{__data:this.getChildren()[0].getOutput(t),__source:this}:{__data:this.getChildren().map(function(e){return e.getOutput(t)}),__source:this}},e.prototype.getData=function(t){return this.dataTransformer(this.getOutput(t))},e.prototype.render=function(t){return this.getChildren().map(function(e){return e.render(t)})},e.prototype.getTag=function(){return"root"},e.prototype.getJsonSchema=function(){var t="id-"+1e9*Math.random(),e=function e(r,n){r&&"object"===i(r)&&(n||r._jsonSchemaCompat!==t)&&(n&&!r._jsonSchemaCompat||(n?delete r._jsonSchemaCompat:r._jsonSchemaCompat=t,n||r.nullable&&r.type&&"object"!==i(r.type)&&(r.type=[r.type,"null"]),Object.keys(r).forEach(function(t){"type"!==t&&e(r[t],n)})))},r=this.getSchema();return e(r,!1),e(r,!0),r},e.prototype.onChildStateChanged=function(t,e,r){this.setState(t)},e.prototype.setState=function(t,e,r){this.setStateSilently(t,e,r),this.getConfig().rootSetState(this.getState(),this,r)},e.prototype.setValue=function(t){this.getChildren().forEach(function(e){return e.setValue(t)})},e}(c);function u(t){return void 0!==t.source}(a=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).default)&&(a.register(s,"NodeType","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\typescript-core\\src\\schemaTypes.tsx"),a.register(c,"Node","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\typescript-core\\src\\schemaTypes.tsx"),a.register(d,"RootNode","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\typescript-core\\src\\schemaTypes.tsx"),a.register(u,"isNodeErrorPure","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\typescript-core\\src\\schemaTypes.tsx")),(l=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).leaveModule)&&l(t)}).call(this,r(4)(t))},function(t,e,r){"use strict";(function(t){var n,o=r(0),i=r(1),s=r(3),a=r.n(s);(n=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).enterModule)&&n(t);"undefined"!=typeof reactHotLoaderGlobal&&reactHotLoaderGlobal.default.signature;var l,c=a.a.div(b||(b=o.__makeTemplateObject(["\n  width: 100%;\n"],["\n  width: 100%;\n"]))),d=a.a.div(g||(g=o.__makeTemplateObject(["\n  font-family: ",";\n  font-size: 1.4vw;\n  display: inline-block;\n"],["\n  font-family: ",";\n  font-size: 1.4vw;\n  display: inline-block;\n"])),function(t){return t.theme.primaryFont}),u=a.a.div(y||(y=o.__makeTemplateObject(["\n  font-family: ",";\n  font-size: 1vw;\n"],["\n  font-family: ",";\n  font-size: 1vw;\n"])),function(t){return t.theme.primaryFont}),f=a.a.div(k||(k=o.__makeTemplateObject(["\n  width: 100%;\n  margin-left: 1vw;\n"],["\n  width: 100%;\n  margin-left: 1vw;\n"]))),p=a.a.div(v||(v=o.__makeTemplateObject(["\n  display: inline-block;\n  float: left;\n  width: 25%;\n"],["\n  display: inline-block;\n  float: left;\n  width: 25%;\n"]))),m=a.a.div(C||(C=o.__makeTemplateObject(["\n  margin-left: 1vw;\n  display: inline-block;\n  width: 70%;\n"],["\n  margin-left: 1vw;\n  display: inline-block;\n  width: 70%;\n"]))),h=a.a.div(L||(L=o.__makeTemplateObject(["\n  display: inline-block;\n  margin-right: 1vw;\n"],["\n  display: inline-block;\n  margin-right: 1vw;\n"])));!function(t){t.NORMAL="normal",t.INLINE_LABEL="inlineLabel"}(l||(l={}));var b,g,y,k,v,C,L,_,j,G=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o.__extends(e,t),e.prototype.render=function(){var t=this.props.parent.getSchema().layout||l.NORMAL,e=this.props.titleExtra?i.createElement(h,null,this.props.titleExtra):null;switch(t){case l.NORMAL:return i.createElement(c,null,i.createElement(d,null,e,this.props.parent.getSchema().title),i.createElement(u,null,this.props.parent.getSchema().description),i.createElement(f,null,this.props.children));case l.INLINE_LABEL:return i.createElement(c,null,i.createElement(p,null,i.createElement(d,null,e,this.props.parent.getSchema().title),i.createElement(u,null,this.props.parent.getSchema().description)),i.createElement(m,null,this.props.children));default:return this.props.children}},e}(i.Component),H=G;e.a=H,(_=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).default)&&(_.register(c,"Wrapper","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithDescription.tsx"),_.register(d,"TitleWrapper","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithDescription.tsx"),_.register(u,"DescriptionWrapper","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithDescription.tsx"),_.register(f,"ContentWrapper","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithDescription.tsx"),_.register(p,"TextWrapperInlineLabel","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithDescription.tsx"),_.register(m,"ContentWrapperInlineLabel","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithDescription.tsx"),_.register(h,"TitleExtraWrapper","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithDescription.tsx"),_.register(l,"LayoutMode","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithDescription.tsx"),_.register(G,"WithDescription","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithDescription.tsx"),_.register(b,"templateObject_1","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithDescription.tsx"),_.register(g,"templateObject_2","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithDescription.tsx"),_.register(y,"templateObject_3","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithDescription.tsx"),_.register(k,"templateObject_4","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithDescription.tsx"),_.register(v,"templateObject_5","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithDescription.tsx"),_.register(C,"templateObject_6","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithDescription.tsx"),_.register(L,"templateObject_7","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithDescription.tsx"),_.register(H,"default","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithDescription.tsx")),(j=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).leaveModule)&&j(t)}).call(this,r(4)(t))},function(t,e,r){"use strict";(function(t){var n,o=r(0),i=r(1),s=r(3),a=r.n(s);(n=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).enterModule)&&n(t);"undefined"!=typeof reactHotLoaderGlobal&&reactHotLoaderGlobal.default.signature;var l,c,d,u,f,p,m,h,b,g=a.a.div(l||(l=o.__makeTemplateObject(["\n    width: 100%;\n"],["\n    width: 100%;\n"]))),y=a.a.div(c||(c=o.__makeTemplateObject(["\n  display: inline-block;\n  height: 3vw;\n  margin-right: 1vw;\n  padding-right: 0.5vw;\n  padding-top: 0.7vw;\n  border-right: 0.15vw solid red;\n"],["\n  display: inline-block;\n  height: 3vw;\n  margin-right: 1vw;\n  padding-right: 0.5vw;\n  padding-top: 0.7vw;\n  border-right: 0.15vw solid red;\n"]))),k=a.a.div(d||(d=o.__makeTemplateObject(["\n  display: inline-block;\n  height: 3vw;\n  margin-right: 1vw;\n  padding-right: 0.5vw;\n  padding-top: 0.7vw;\n  border-right: 0.15vw solid transparent;\n"],["\n  display: inline-block;\n  height: 3vw;\n  margin-right: 1vw;\n  padding-right: 0.5vw;\n  padding-top: 0.7vw;\n  border-right: 0.15vw solid transparent;\n"]))),v=a.a.div(u||(u=o.__makeTemplateObject(["\n  color: red;\n  opacity: ",";\n"],["\n  color: red;\n  opacity: ",";\n"])),function(t){return t.isVisible?1:0}),C=a.a.div(f||(f=o.__makeTemplateObject(["\n  display: inline-block;\n  width: 80%;\n"],["\n  display: inline-block;\n  width: 80%;\n"]))),L=a.a.span(p||(p=o.__makeTemplateObject(["\n    margin-left: 0.3vw;\n    font-weight: bold;\n"],["\n    margin-left: 0.3vw;\n    font-weight: bold;\n"]))),_=a.a.div(m||(m=o.__makeTemplateObject(["\n    max-width: 10vw;\n"],["\n    max-width: 10vw;\n"]))),j=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o.__extends(e,t),e.prototype.render=function(){var t=this.props.context.getErrorsForNode(this.props.parent)||[],e=null;return t.length>0&&(e=i.createElement("div",null,i.createElement(_,null,i.createElement("div",null,i.createElement(L,null,"Error!")),t[0].message))),i.createElement(g,null,0===t.length?i.createElement(k,null,e):i.createElement(y,null,e),i.createElement(C,null,this.props.children))},e}(i.Component),G=j;e.a=G,(h=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).default)&&(h.register(g,"Wrapper","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithErrors.tsx"),h.register(y,"ErrorIcon","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithErrors.tsx"),h.register(k,"NoErrorIcon","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithErrors.tsx"),h.register(v,"IconWrapper","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithErrors.tsx"),h.register(C,"Content","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithErrors.tsx"),h.register(L,"ErrorPopoverLabel","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithErrors.tsx"),h.register(_,"ErrorContainer","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithErrors.tsx"),h.register(j,"WithErrors","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithErrors.tsx"),h.register(l,"templateObject_1","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithErrors.tsx"),h.register(c,"templateObject_2","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithErrors.tsx"),h.register(d,"templateObject_3","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithErrors.tsx"),h.register(u,"templateObject_4","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithErrors.tsx"),h.register(f,"templateObject_5","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithErrors.tsx"),h.register(p,"templateObject_6","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithErrors.tsx"),h.register(m,"templateObject_7","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithErrors.tsx"),h.register(G,"default","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithErrors.tsx")),(b=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).leaveModule)&&b(t)}).call(this,r(4)(t))},function(t,e,r){"use strict";(function(t){var n,o=r(0),i=r(1),s=r(3),a=r.n(s);(n=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).enterModule)&&n(t);"undefined"!=typeof reactHotLoaderGlobal&&reactHotLoaderGlobal.default.signature;var l,c,d,u=a.a.div(l||(l=o.__makeTemplateObject(["\n    margin-top: 1vw;\n    width: 100%;\n"],["\n    margin-top: 1vw;\n    width: 100%;\n"]))),f=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o.__extends(e,t),e.prototype.render=function(){return i.createElement(u,null,this.props.children)},e}(i.Component),p=f;e.a=p,(c=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).default)&&(c.register(u,"Wrapper","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithMargins.tsx"),c.register(f,"WithMargins","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithMargins.tsx"),c.register(l,"templateObject_1","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithMargins.tsx"),c.register(p,"default","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\utils\\WithMargins.tsx")),(d=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).leaveModule)&&d(t)}).call(this,r(4)(t))},function(t,e,r){"use strict";(function(t){r.d(e,"a",function(){return l});var n,o=r(0),i=r(5);(n=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).enterModule)&&n(t);"undefined"!=typeof reactHotLoaderGlobal&&reactHotLoaderGlobal.default.signature;var s,a,l=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o.__extends(e,t),e.prototype.resolveInitialState=function(){return{value:this.getInitialValue()}},e.prototype.getRawOutput=function(){return this.getState().value},e.prototype.render=function(t){return this.renderSimple(this.getState().value,t)},e}(i.a);(s=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).default)&&s.register(l,"SimpleNode","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\typescript-core\\src\\simpleNodes.tsx"),(a=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).leaveModule)&&a(t)}).call(this,r(4)(t))},function(t,e,r){"use strict";(function(t){var n,o=r(12),i=r(18),s=r(19),a=r(20),l=r(21),c=r(22);(n=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).enterModule)&&n(t);"undefined"!=typeof reactHotLoaderGlobal&&reactHotLoaderGlobal.default.signature;var d,u,f={ROOT:{default:o.a},STRING:{default:i.a,notice:s.a,enum:a.a},OBJECT:{default:o.a},ARRAY:{default:l.a,tuple:c.a}};e.a=f,(d=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).default)&&d.register(f,"default","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\defaultReactHandlerProvider.tsx"),(u=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).leaveModule)&&u(t)}).call(this,r(4)(t))},function(t,e,r){"use strict";(function(t){r.d(e,"a",function(){return l});var n,o=r(0),i=r(5);(n=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).enterModule)&&n(t);"undefined"!=typeof reactHotLoaderGlobal&&reactHotLoaderGlobal.default.signature;var s,a,l=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o.__extends(e,t),e.prototype.renderComposite=function(t,e){return Object.keys(e).map(function(t){return e[t]})},e.prototype.resolveInitialState=function(){var t={};return Object.keys(this.getChildrenMapFromSchema()).forEach(function(e){t[e]=null}),t},e.prototype.getValueMapFromValue=function(t){return o.__assign({},t)},e.prototype.setValue=function(t){var e=this;this.getSchema().formatInput&&(t=this.getSchema().formatInput(t)),t=this.getValueMapFromValue(t),Object.keys(this.getChildrenMapFromSchema()).forEach(function(r){e.findChild(r)&&e.findChild(r).setValue(t[r])})},e.prototype.getRawOutput=function(t){var e=this,r={};return Object.keys(this.getChildrenMapFromSchema()).forEach(function(n){e.findChild(n)&&e.findChild(n).isOutputAvailable()&&(r[n]=e.findChild(n).getOutput(t))}),this.getCompositeOutput(r)},e.prototype.onChildStateChanged=function(t,e,r){var n;this.setState(((n={})[e.getTag()]=t,n))},e.prototype.resolveChildren=function(){var t=this,e=this.getChildrenMapFromSchema();return Object.keys(e).map(function(r){var n=t.resolveNode(e[r],t,t.getConfig());return n.setTag(r),n})},e.prototype.render=function(t){var e=this,r={};return Object.keys(this.getChildrenMapFromSchema()).forEach(function(n,o){e.findChild(n)&&(r[n]=e.findChild(n).render(t))}),this.renderComposite(t,r)},e}(i.a);(s=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).default)&&s.register(l,"CompositeNode","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\typescript-core\\src\\compositeNodes.tsx"),(a=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).leaveModule)&&a(t)}).call(this,r(4)(t))},function(t,e,r){"use strict";(function(t){var n,o=r(0),i=r(11);(n=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).enterModule)&&n(t);"undefined"!=typeof reactHotLoaderGlobal&&reactHotLoaderGlobal.default.signature;var s,a,l=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o.__extends(e,t),e.prototype.getChildrenMapFromSchema=function(){return this.getSchema().properties},e.prototype.getCompositeOutput=function(t){return t},e}(i.a),c=l;e.a=c,(s=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).default)&&(s.register(l,"ObjectDefault","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\ObjectDefault.tsx"),s.register(c,"default","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\ObjectDefault.tsx")),(a=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).leaveModule)&&a(t)}).call(this,r(4)(t))},function(t,e,r){"use strict";(function(t){r.d(e,"a",function(){return m});var n,o=r(0),i=r(1),s=r(3),a=r.n(s),l=r(14),c=r(10);(n=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).enterModule)&&n(t);"undefined"!=typeof reactHotLoaderGlobal&&reactHotLoaderGlobal.default.signature;var d,u,f,p=a.a.div(d||(d=o.__makeTemplateObject(["\n  width: 100%;\n"],["\n  width: 100%;\n"]))),m=function(t){function e(e){var r=t.call(this,e)||this;return r.client=new l.a,r.client.configure({rootSetState:function(t,e){return r.handleFormStateUpdate(t,e)},rootSetContext:function(t,e){return r.handleFormContextUpdate(t,e)},rootModifyContext:function(t,e){return r.handleFormContextMapping(t,e)},rootState:{}}),r.props.use&&r.props.use.forEach(function(t){r.client.registerHandlers(t)}),r.client.registerHandlers(c.a),r.state={tree:null,formContext:{errors:[],getErrorsForNode:function(){return[]}}},r}return o.__extends(e,t),e.prototype.handleFormStateUpdate=function(t,e){var r=this;this.setState({tree:e},function(){!1!==r.props.validateOnChange&&r.state.tree.validate()})},e.prototype.handleFormContextUpdate=function(t,e){this.setState({formContext:o.__assign({},this.state.formContext,t)})},e.prototype.handleFormContextMapping=function(t,e){this.handleFormContextUpdate(t(this.state.formContext),e)},e.prototype.createTree=function(){var t=this,e=this.props.children;this.state.tree||setTimeout(function(){t.client.render(e);var r=t.client.getTree();t.setState({tree:r},function(){!1!==t.props.validateOnInit&&t.state.tree.validate()})})},e.prototype.componentDidMount=function(){this.createTree()},e.prototype.render=function(){var t=this;return this.createTree(),i.createElement(p,null,this.state.tree?this.state.tree.render(this.state.formContext):null,this.state.tree&&this.props.onSubmit?i.createElement("div",{style:{marginLeft:"-2vw"}},i.createElement("button",{onClick:function(){t.state.tree&&t.props.onSubmit&&t.props.onSubmit(t.state.tree.getData())}},"Save")):null)},e}(i.Component);(u=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).default)&&(u.register(p,"Wrapper","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\form.tsx"),u.register(m,"Form","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\form.tsx"),u.register(d,"templateObject_1","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\form.tsx")),(f=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).leaveModule)&&f(t)}).call(this,r(4)(t))},function(t,e,r){"use strict";(function(t){var n,o=r(0),i=r(15);(n=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).enterModule)&&n(t);"undefined"!=typeof reactHotLoaderGlobal&&reactHotLoaderGlobal.default.signature;var s,a,l=function(){function t(){this.tree=null,this.state=null,this.stateTransformers=[],this.contextUpdateHandlers=[],this.contextTransformers=[],this.rootConfig={}}return t.prototype.configure=function(t){this.rootConfig=o.__assign({},this.rootConfig,t)},t.prototype.registerHandlers=function(t){var e=o.__assign({},this.rootConfig.handlers);Object.keys(t).forEach(function(r){e[r]=o.__assign({},e[r],t[r])}),this.rootConfig=o.__assign({},this.rootConfig,{handlers:o.__assign({},this.rootConfig.handlers,e)})},t.prototype.handleFormStateUpdate=function(t,e){var r=this;this.state=t,this.stateTransformers.forEach(function(t){r.state=t(r.state,e)})},t.prototype.handleFormContextUpdate=function(t,e){this.contextUpdateHandlers.forEach(function(r){r(t,e)})},t.prototype.handleFormContextMapping=function(t,e){this.contextTransformers.forEach(function(t){t(context,e)})},t.prototype.render=function(t,e){var r=this;return void 0===e&&(e=null),this.tree=Object(i.a)(t,null,o.__assign({rootSetState:function(t,e){return r.handleFormStateUpdate(t,e)},rootSetContext:function(t,e){return r.handleFormContextUpdate(t,e)},rootModifyContext:function(t,e){return r.handleFormContextMapping(t,e)},rootState:{}},this.rootConfig,e)),this},t.prototype.getTree=function(){return this.tree},t.prototype.getState=function(){return this.state},t}(),c=l;e.a=c,(s=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).default)&&(s.register(l,"BlinkformsClient","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\typescript-core\\src\\BlinkformsClient.tsx"),s.register(c,"default","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\typescript-core\\src\\BlinkformsClient.tsx")),(a=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).leaveModule)&&a(t)}).call(this,r(4)(t))},function(t,e,r){"use strict";(function(t){r.d(e,"a",function(){return C});var n,o=r(0),i=r(5),s=r(16),a=r.n(s),l=r(17);(n=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).enterModule)&&n(t);var c,d;"undefined"!=typeof reactHotLoaderGlobal&&reactHotLoaderGlobal.default.signature;function u(t,e){return t.ui?e[t.ui]:e.default}function f(t,e){switch(t.type){case i.b.OBJECT:return e.handlers.OBJECT;case i.b.STRING:return e.handlers.STRING;case i.b.NUMBER:return e.handlers.NUMBER;case i.b.BOOLEAN:return e.handlers.BOOLEAN;case i.b.ARRAY:return e.handlers.ARRAY;case i.b.ROOT:return null;default:throw"Unknown node type "+t.type}}function p(t,e,r,n){var o=new n(n,t.type,t,e,r,m);return o.resolve(),e.addChild(o),o}function m(t,e,r){return p(t,e,r,u(t,f(t,r)))}function h(t){return t&&void 0!==t.__data&&t.__source}function b(t,e){return h(t)?null!==e&&0===e.length?t.__source:b(t.__data,e):null===e?t:"string"==typeof e?b(t,e.split(".")):0===e.length?b(t,null):b(t[e[0]],e.slice(1))}function g(t,e){for(var r=[],n=void 0,o=/^([^\.\[]+)|\.([^\.\[]+)|\["([^"]+)"\]|\[(\d+)\]/g;n=o.exec(e);)r.push(Array.from(n).slice(1).filter(function(t){return void 0!==t})[0]);return b(t,r)}function y(t){if(h(t))return y(t.__data);if(Array.isArray(t))return t.map(function(t){return y(t)});if(t instanceof Object){var e={};return Object.keys(t).forEach(function(r){e[r]=y(t[r])}),e}return t}function k(t,e,r){var n;return function(){var o=this,i=arguments,s=r&&!n;clearTimeout(n),n=setTimeout(function(){n=null,r||t.apply(o,i)},e),s&&t.apply(o,i)}}function v(t){return o.__awaiter(this,void 0,void 0,function(){var e,r,n,s,l,c,d;return o.__generator(this,function(u){return e=new a.a(o.__assign({allErrors:!0},t.getConfig().ajvOptions)),r=e.compile(t.getJsonSchema()),n=t.getOutput({enableFormat:!1}),s=y(n),r(s),l=r.errors||[],c=new WeakMap,d=t.validateCustom(),(l=l.concat(d)).forEach(function(t){var e=null;if(null!==(e=Object(i.d)(t)?t.source:g(n,t.dataPath))){var r=t;c.has(e)?c.get(e).push(r):c.set(e,[r])}}),t.setContext({errors:l||[],getErrorsForNode:function(t){return c.get(t)||[]}}),[2]})})}function C(t,e,r){void 0===e&&(e=null),void 0===r&&(r=null);var n=o.__assign({},l.a,r);return n.uidGenerator||(n.uidGenerator=n.uidGeneratorFactory()),e||(e=new i.c(t,n,m,y,k(v,750))).resolve(),m(t,e,n),e}(c=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).default)&&(c.register(u,"getHandlerForUI","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\typescript-core\\src\\schemaParser.tsx"),c.register(f,"getHandlerForType","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\typescript-core\\src\\schemaParser.tsx"),c.register(p,"createNode","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\typescript-core\\src\\schemaParser.tsx"),c.register(m,"recTransformSchemaIntoTree","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\typescript-core\\src\\schemaParser.tsx"),c.register(h,"isNodeMetaOutputValue","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\typescript-core\\src\\schemaParser.tsx"),c.register(b,"recGetMetaOutputSourceNodeByPath","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\typescript-core\\src\\schemaParser.tsx"),c.register(g,"getMetaOutputSourceNodeByPath","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\typescript-core\\src\\schemaParser.tsx"),c.register(y,"transformOutputToRawData","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\typescript-core\\src\\schemaParser.tsx"),c.register(k,"debounce","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\typescript-core\\src\\schemaParser.tsx"),c.register(v,"validateRoot","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\typescript-core\\src\\schemaParser.tsx"),c.register(C,"transformSchemaIntoTree","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\typescript-core\\src\\schemaParser.tsx")),(d=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).leaveModule)&&d(t)}).call(this,r(4)(t))},function(t,e){t.exports=require("ajv")},function(t,e,r){"use strict";(function(t){var n;r.d(e,"a",function(){return s}),(n=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).enterModule)&&n(t);"undefined"!=typeof reactHotLoaderGlobal&&reactHotLoaderGlobal.default.signature;var o,i,s={handlers:{ROOT:{default:null},STRING:{default:null},OBJECT:{default:null},NUMBER:{default:null},BOOLEAN:{default:null},ARRAY:{default:null}},rootState:null,rootSetState:function(){},rootSetContext:function(){},rootModifyContext:function(){},uidGenerator:null,uidGeneratorFactory:function(){var t=0;return function(){return++t}}};(o=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).default)&&o.register(s,"defaultParserConfig","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\typescript-core\\src\\defaultParserConfig.tsx"),(i=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).leaveModule)&&i(t)}).call(this,r(4)(t))},function(t,e,r){"use strict";(function(t){var n,o=r(0),i=r(1),s=r(3),a=r.n(s),l=r(6),c=r(7),d=r(8),u=r(9);(n=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).enterModule)&&n(t);"undefined"!=typeof reactHotLoaderGlobal&&reactHotLoaderGlobal.default.signature;var f,p,m,h=a.a.div(f||(f=o.__makeTemplateObject(["\n  width: 10vw;\n  margin-right: 3vw;\n"],["\n  width: 10vw;\n  margin-right: 3vw;\n"]))),b=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o.__extends(e,t),e.prototype.getInitialValue=function(){return""},e.prototype.renderSimple=function(t,e){var r=this;return i.createElement(d.a,{parent:this},i.createElement(c.a,{parent:this,context:e},i.createElement(l.a,{parent:this},i.createElement(h,null,i.createElement("input",{value:t,onChange:function(t){r.setState({value:t.target.value})}})))))},e}(u.a),g=b;e.a=g,(p=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).default)&&(p.register(h,"InputWrapper","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\StringDefault.tsx"),p.register(b,"StringDefault","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\StringDefault.tsx"),p.register(f,"templateObject_1","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\StringDefault.tsx"),p.register(g,"default","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\StringDefault.tsx")),(m=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).leaveModule)&&m(t)}).call(this,r(4)(t))},function(t,e,r){"use strict";(function(t){var n,o=r(0),i=r(1),s=r(9);(n=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).enterModule)&&n(t);"undefined"!=typeof reactHotLoaderGlobal&&reactHotLoaderGlobal.default.signature;var a,l,c=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o.__extends(e,t),e.prototype.getInitialValue=function(){return"null"},e.prototype.isOutputAvailable=function(){return!1},e.prototype.validateCustom=function(){return[]},e.prototype.renderSimple=function(t,e){return i.createElement("div",null,this.getSchema().value)},e}(s.a),d=c;e.a=d,(a=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).default)&&(a.register(c,"StringNotice","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\StringNotice.tsx"),a.register(d,"default","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\StringNotice.tsx")),(l=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).leaveModule)&&l(t)}).call(this,r(4)(t))},function(t,e,r){"use strict";(function(t){var n,o=r(0),i=r(1),s=r(6),a=r(7),l=r(8),c=r(9);(n=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).enterModule)&&n(t);"undefined"!=typeof reactHotLoaderGlobal&&reactHotLoaderGlobal.default.signature;var d,u,f=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o.__extends(e,t),e.prototype.getInitialValue=function(){return""},e.prototype.renderSimple=function(t,e){var r=this,n=this.getSchema().enumLabels||null,o=(this.getSchema().enum||[]).map(function(t){var e=t,r=t;if(n){var o=n.find(function(e){return e.name===t});o&&(e=o.label,r=o.name||t)}return{name:r,label:e}}),c=o.find(function(e){return e.name===t})||null;return i.createElement(l.a,{parent:this},i.createElement(a.a,{parent:this,context:e},i.createElement(s.a,{parent:this},i.createElement("select",{onChange:function(t){return r.setState({value:t.target.value})}},o.map(function(t,e){return i.createElement("option",{key:"opt-"+t.name+"-"+e,value:t.name,selected:c===t.name&&null!==t.name&&void 0!==t.name},t.label)})))))},e}(c.a),p=f;e.a=p,(d=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).default)&&(d.register(f,"StringEnum","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\StringEnum.tsx"),d.register(p,"default","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\StringEnum.tsx")),(u=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).leaveModule)&&u(t)}).call(this,r(4)(t))},function(t,e,r){"use strict";(function(t){var n,o=r(0),i=r(1),s=r(6),a=r(7),l=r(8),c=r(5);(n=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).enterModule)&&n(t);"undefined"!=typeof reactHotLoaderGlobal&&reactHotLoaderGlobal.default.signature;var d,u,f=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o.__extends(e,t),e.prototype.resolveInitialState=function(){return{items:{}}},e.prototype.setValue=function(t){var e=this;this.getSchema().formatInput&&(t=this.getSchema().formatInput(t)),t=o.__assign({},t),Object.keys(t).length!==this.getChildren().length&&this.overrideChildren(Object.keys(t).map(function(t){var r=e.resolveNode(e.getSchema().items,e,e.getConfig());return r.setTag(t),r})),this.getChildren().forEach(function(e,r){e.setValue(t[r])})},e.prototype.getRawOutput=function(t){var e=[];return this.getChildren().forEach(function(r){r.isOutputAvailable()&&e.push(r.getOutput(t))}),e},e.prototype.onChildStateChanged=function(t,e,r){var n;this.setState({items:o.__assign({},this.getState().items,(n={},n[e.getTag()]=t,n))})},e.prototype.removeItem=function(t){var e=o.__assign({},this.getState().items);delete e[t],this.overrideChildren(this.getChildren().filter(function(e,r){return t!==r}).map(function(t,e){return t.setTag(e.toString()),t})),this.setState({items:e})},e.prototype.addNewItem=function(){var t,e=this.getChildren().length,r=this.resolveNode(this.getSchema().items,this,this.getConfig());this.setState({items:o.__assign({},this.getState().items,(t={},t[e]=r.resolveInitialState(),t))})},e.prototype.resolveChildren=function(){return[]},e.prototype.render=function(t){var e=this;return i.createElement(l.a,{parent:this},i.createElement(a.a,{parent:this,context:t},i.createElement(s.a,{parent:this,titleExtra:i.createElement("div",null,i.createElement("button",{onClick:function(){return e.addNewItem()}},"Add item"))},this.getChildren().map(function(r,n){return i.createElement("div",null,i.createElement("div",null,i.createElement("button",{onClick:function(){return e.removeItem(n)}},"X")),r.render(t))}))))},e}(c.a),p=f;e.a=p,(d=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).default)&&(d.register(f,"ArrayDefault","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\ArrayDefault.tsx"),d.register(p,"default","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\ArrayDefault.tsx")),(u=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).leaveModule)&&u(t)}).call(this,r(4)(t))},function(t,e,r){"use strict";(function(t){var n,o=r(0),i=r(1),s=r(11),a=r(6),l=r(7),c=r(8);(n=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).enterModule)&&n(t);"undefined"!=typeof reactHotLoaderGlobal&&reactHotLoaderGlobal.default.signature;var d,u,f=function(t){function e(){return null!==t&&t.apply(this,arguments)||this}return o.__extends(e,t),e.prototype.getChildrenMapFromSchema=function(){return o.__assign({},this.getSchema().items)},e.prototype.setValue=function(t){var e=this;this.getSchema().formatInput&&(t=this.getSchema().formatInput(t)),t=this.getValueMapFromValue(t),Object.keys(this.getChildrenMapFromSchema()).forEach(function(r){e.findChild(r)&&e.findChild(r).setValue(t[r])})},e.prototype.getCompositeOutput=function(t){var e=[];return Object.keys(t).sort().forEach(function(r){return e.push(t[r])}),e},e.prototype.renderComposite=function(t,e){return i.createElement(c.a,{parent:this},i.createElement(l.a,{parent:this,context:t},i.createElement(a.a,{parent:this},Object.keys(e).map(function(t){return e[t]}))))},e}(s.a),p=f;e.a=p,(d=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).default)&&(d.register(f,"ArrayTuple","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\ArrayTuple.tsx"),d.register(p,"default","C:\\Users\\Piotr\\gitdir\\blinkforms-docs\\blinkforms\\react-json-blinkforms\\src\\renderers\\ArrayTuple.tsx")),(u=("undefined"!=typeof reactHotLoaderGlobal?reactHotLoaderGlobal:r(2)).leaveModule)&&u(t)}).call(this,r(4)(t))},function(t,e,r){"use strict";r.r(e);var n=r(13);r.d(e,"Form",function(){return n.a});r(10),"undefined"!=typeof reactHotLoaderGlobal&&reactHotLoaderGlobal.default.signature}]));
+//# sourceMappingURL=index.js.map
